If a seq has already been transformed, return immediately

+++ app.py
@@ -1,11 +1,10 @@
 #!flask/bin/python
 from flask import Flask, request, jsonify, render_template
-from bokeh.palettes import Category10
 from squiggle import transform
 from skbio.io import read
 import pandas as pd
 import xxhash
-
+import os.path
 
 app = Flask(__name__)
 
@@ -21,14 +20,12 @@ def seq_query():
     zone = df.loc[(float(request.args.get("x_max", df.x.max())) >= df.x) &
                   (float(request.args.get("x_min", df.x.min())) <= df.x)].values
 
-    # seq_number = int(request.args.get("seq_number"))
-    # if seq_number > len(Category10):
-    #     seq_color = Category10[3][0]
-    # else:
-    #     seq_color = Category10[seq_number + 1 if seq_number > 2 else 3][seq_number]
+    zone = zone.tolist()
+    if len(zone) > 5000:
+        zone = zone[::int(len(zone) / 5000)]
 
     return jsonify({"name": request.args["seq_id"],
-                    "data": zone.tolist(),
+                    "data": zone,
                     "marker": False})
 
 @app.route("/fasta", methods=["POST"])
@@ -38,11 +35,11 @@ def parse_fasta():
     for seq in read([x.decode("ascii") for x in request.files["sequence"].readlines()], "fasta"):
         transformed = list(zip(*transform(str(seq))))
         seq_hash = str(xxhash.xxh64(str(seq)).intdigest())
-        with open("data/" + seq_hash + ".csv", "w+") as f:
-            f.write("x,y\n")
-            for coord in transformed:
-                f.write("%f,%f\n" % (coord[0], coord[1]))
-
+        if not os.path.exists("data/" + seq_hash + ".csv"):
+            with open("data/" + seq_hash + ".csv", "w+") as f:
+                f.write("x,y\n")
+                for coord in transformed:
+                    f.write("%f,%f\n" % (coord[0], coord[1]))
         results.append(dict(seq_hash=seq_hash, seq_id=seq.metadata["id"]))
     return jsonify(results)
